+++ tests/test_infer.py
@@ -6,8 +6,8 @@ import pytest
 import datacraft
 from datacraft.infer import from_examples, csv_to_spec
 
-from datacraft._infer.helpers import (_all_is_numeric, _calculate_weights, _are_values_unique)
-from datacraft._infer.num_analyzers import _compute_range
+from datacraft._infer.helpers import (all_is_numeric, calculate_weights, are_values_unique)
+from datacraft._infer.num_analyzers import compute_range
 
 from .test_utils import deep_sort
 
@@ -76,37 +76,37 @@ EXAMPLES = [
 
 
 def test_is_numeric():
-    assert _all_is_numeric([1, 2, 3])
-    assert not _all_is_numeric([1, 2, '3'])
-    assert not _all_is_numeric([True, 1, 2])
+    assert all_is_numeric([1, 2, 3])
+    assert not all_is_numeric([1, 2, '3'])
+    assert not all_is_numeric([True, 1, 2])
 
 
 def test_calculate_weights():
     values = ["a", "b", "c", "a", "a", "b", "d", "a", "a", "a"]
     expected = {"a": 0.6, "b": 0.2, "c": 0.1, "d": 0.1}
-    assert _calculate_weights(values) == expected
+    assert calculate_weights(values) == expected
 
 
 def test_are_values_unique():
-    assert _are_values_unique([1, 2, 3, 4])
-    assert not _are_values_unique([1, 2, 3, 4, 4])
+    assert are_values_unique([1, 2, 3, 4])
+    assert not are_values_unique([1, 2, 3, 4, 4])
 
 
 def test_compute_int_range():
     values = [1, 5, 3, 4, 2]
     expected = {"type": "rand_int_range", "data": [1, 5]}
-    assert _compute_range(values) == expected
+    assert compute_range(values) == expected
 
 
 def test_invalid_data_in_range():
     with pytest.raises(ValueError):
-        _compute_range([1, 2, 3, 'a'])
+        compute_range([1, 2, 3, 'a'])
 
 
 def test_compute_float_range():
     values = [1.1, 5.1, 3.1, 4.1, 2.1]
     expected = {"type": "rand_range", "data": [1.1, 5.1]}
-    assert _compute_range(values) == expected
+    assert compute_range(values) == expected
 
 
 def test_from_example_simple():